Configure the app based on Heroku settings
I know I'm going to be running this on Heroku for the moment, so I might
as well get that started here.

By default, this library forces SSL database connections. I've added a
conditional to override this and disable SSL-forcing in development,
because I really don't want to go down the rabbit hole of provisioning a
self-signed SSL certificate for my own development database.

I could have gone with the option of a local SQLite database instead of
going full Postgres, but I remember from past experience that Postgres
and SQLite differ in some interesting ways, and I wouldn't want that to
bite me when deploy time comes around.