Add a first-pass "shop" view
This is the view for planning out a regular shopping trip. For now, all
it does it let us put together a list of meals and lists out the
ingredients (and delete meals that we decide we don't want). Next up is
to assign them to a batch object, and then automatically set up our
grocery list.

I added a custom manager object to keep too much model logic from
creeping into the view layer. This seems especially useful as the state
machine (from suggested to planned to completed or cancelled) isn't
immediately apparent from reading the fields involved.