Add a first version of an ingredient form
The plan is that we're going to use this to go through the kitchen
checking off what we have and what we need. For now, the data on stuff
that we already have is getting thrown away. We could keep a record of
it later if we find a use for it, but for now it's only for the benefit
of the human filling out the form.