Update our grocery list automatically
Todoist currently has an API library, so we can use that with a
relatively small wrapper (injecting configuration from Django's settings
so that I can theoretically write tests against it if and when we start
wanting to add new features).

The UI is well and truly basic, and the form validation is pretty much
non-existent (because there's currently only one user of this thing,
with hopefully a second one coming soon, but that's it).