Rename "Plan" and "Shop" to "Now" and "Next"
In the original use, "Plan" is intended as a noun (this is the meal
plan) and "Shop" is intended as a verb (let's go shopping). This is
pretty confusing to someone trying to figure out what's going on. "Now"
and "Next" are at least both the same type of word, so that should make
things a little clearer.