Wrap the `render` shortcut in some instrumentation
Hopefully this will be useful in showing when template rendering is
slowing requests down (if that ever even happens). It's also useful for
showing when database queries are being made *during* template
rendering, which is probably fine for one-off queries, but if it's
happening a lot, it's probably worth making sure the queries execute
up-front.