Explicitly exclude ticked-off groceries
It looks like Todoist's API might have changed here: the `items.all`
method now seems to return completed items as well as outstanding ones.
This breaks the app's functionality, as anything that we've previously
added to our grocery list will not be re-added, even if we've since
checked it off the list.

This change explicitly checks an item's `checked` field, to make sure
it's still outstanding.

If Todoist happen to revert this change (they were previously touting
completed items as a "premium feature", so this might be a bug), the
extra check won't do any harm.