Fix a bug when adding items to the grocery list
When we add items, we weren't populating the `checked` field, and
letting it default to False. Because we're now inspecting the `checked`
field to filter out completed items, and because the Todoist API doesn't
have this field's default baked into its model, this was causing a
`KeyError`.

In normal Python, I'd use something like:

    if not item.get("checked", False)

...but Todoist's `Item` model doesn't provide a `get` method. Rather
than add a `try...except`, or an extra `checked in item` condition, it's
easier to populate the `checked` field explicitly and make all the items
look the same.