Migrate ingredients to their own model
It was handy to have them in an ArrayField for a while, but in order to
do handy things like different ingredient categories (for separating out
fruit and veg, say), we're going to want them in their own model.

I maybe went a bit overboard on the migrations (adding a separate
reversible data migration), but I wanted to experiment a bit.