Handle ingredient shopping with a Django form
The main benefit of this is that there's some validation that all the
ingredients have a "Got it" or "Need it" option selected, avoiding the
problem where everything is treated as not needed by default, so that
someone tapping on the "Add to order" button accidentally adds nothing
to the grocery list.

It also means there's some shared logic between the page that displays
the form and the page that interprets it, which means if we want to
change things like the form naming convention, we only need to do it in
one place.

Also, tests.