Make the grocery list use Django's cache
The Todoist Python library does support caching by default, but it tries
to cache to a file on disk, rather than using a pluggable cache layer.
So I've had a crack at extending it to plug into Django's cache
framework (or anything with the same `get`/`set` API).

I've added a bit of instrumentation to see whether it makes any
difference to performance. Early indications suggest it might not, but
let's see what it does once it's deployed.