Don't expire cached grocery list state
Because we only use this to get a head-start on syncing, rather than
relying on it being up to date, we can afford not to have it expire at
all. Also, because we only go through this process once a week or so, we
don't get a whole lot of benefit from only caching data for the default
of 5 minutes.