Skip updating the cache if it's not out of date
The cache here is pretty large (on the order of tens of kilobytes), so
we save a wodge of time by not updating it when we don't have to.

We're using the sync token here as a proxy for the state being
unchanged, because that appears to be how the Todoist API works.