Allow marking dishes as inactive
This means we get to remove them from our list of candidates when we're
putting an order together, but they're still in the historical record of
meals we've eaten. Also, if we happen to deactivate a dish while it's in
one or both of our lists, it'll still show up.