Add friendlier date names for scheduled meals
Because we're usually dealing with dates in a well-defined range (the
next week or so), we should be able to get away with much shorter names
than we would be able to get away with in a less constrained scenario.

I've kept the `date_name` function separate from the template tag,
because I know I'm going to be using it in a form next, so it makes
sense to keep it separate from the outset.