Log active/inactive flag states on every request
For the moment, I'm going with a relatively blunt approach of loading up
all the flag that exist and recording whether they're active. At some
point, this might become impractical, but I suspect I'll never have that
many flags existing at once.

I've had to add a separate memory-based cache for the feature flags, due
to the fact that I'm using a database-backed cache for the GroceryList
class. When Waffle tries to use this, it ends up running a bunch of
queries against the database cache to avoid running a single query
against the database itself. Which is a little silly.