Add a separate view for date planning
I *think* this broadly makes sense. Right now, I expect that we're only
going to use explicit date-planning after we've ordered each batch, so
it seems better to keep it on a separate screen rather than using up
valuable area on the main screen.

We'll see how well that works out, but this is good enough for a start.