Tidy up the date-mode-specific styles
Instead of having three different styles -- one for the "Next" page, one
for the "Now" page with dates, one for the "Now" page without dates --
we now only have one style for each page, so we can clean up some of the
style rules.

There might be some scope in future for taking that down to one style,
but for the moment we have significantly more information and more
actions to display on the "Now" page.