Build available meal dates at form-building time
There was a bug in this code that meant the set of available dates (and
their corresponding descriptions) was getting baked into the
MealDateForm at module-loading time, rather than each time the form got
built. Now that this is running on a paid plan that sometime keeps the
instance up and running overnight, this had the unfortunate effect of
throwing the dates off for the date-planning forms.

To fix this, I'm building up the set of choices for the `date` field's
widget within the `__init__` method, rather than defining them on the
class. One nice benefit of this is that it's now possible to pass the
current meal's planned date into the `week_choices` function, meaning
that the complexity of checking whether that's in the default set of
choices falls out and disappears. In short, instead of saying:

    Is <date> in the default set? No? Add it and re-sort the list.

we get to say:

    Build me a set of choices, and make sure to include <date>.

Another nice side-effect of this change is that the logic for mapping
dates onto Django's (value, label) choice tuples is now only getting
called in one place, so the `choice_for_date` function can get inlined
into `week_choices`.