Add the 'apple-mobile-web-app-capable' meta tag
This should let the site operate as its own separate app on iOS, rather
than opening up in Safari every time.

Hiding this behind a feature flag for the moment, because I don't have a
great deal of faith in my understanding of how web stuff works on iOS,
so I don't want to break things.