Migrate dependency management to Pipenv
This gives better control over dependency versions than straight `pip`,
because it's possible to specify version constraints (in `Pipfile`)
separately from the current dependency versions (in `Pipfile.lock`).

I held off moving to this at the start of the project, partly to keep the
complexity to a minimum and partly so I didn't have to decide between
Pipenv and Poetry, but now it's time to get with the times.

I have no strong preference for either tool, but it looks like there are
more tools to migrate from Pipenv to Poetry than there are the other
way, so it seems like the more easily reversible choice. Also, Heroku
claims to support Pipenv natively, and doesn't support Poetry, so that's
another layer of complexity I don't need to think about.