Tweak the way we order ingredient storage
This originally went with reverse alphabetical order, so that ungrouped
items would turn up at the end of the list. This also meant that the
labels were in reverse alphabetical order, which bugged me more than I
thought it would. So now there's a bit of OrderedDict magic to keep
things in forward alphabetical order, but to explicitly put ungrouped
items at the end of the list. Which is a more explicit way of doing
things anyway, rather than having it turn up as a by-product of the
query sorting.